Add Member parent class
